+++ lib/pmdtester/project.rb
@@ -1,3 +1,5 @@
+require_relative './pmd_branch_detail'
+
 module PmdTester
   # This class represents all the information about the project
   class Project
@@ -9,7 +11,7 @@ module PmdTester
     attr_reader :webview_url
     attr_reader :tag
     attr_reader :exclude_pattern
-    attr_accessor :diffs_exist
+    attr_accessor :report_diff
     # key: pmd branch name as String => value: local path of pmd report
 
     def initialize(project)
@@ -17,16 +19,28 @@ module PmdTester
       @type = project.at_xpath('type').text
       @connection = project.at_xpath('connection').text
 
-      webview_url_element = project.at_xpath('webview-url')
-      @webview_url = @connection
-      @webview_url = webview_url_element.text unless webview_url_element.nil?
-
+      @tag = 'master'
       tag_element = project.at_xpath('tag')
       @tag = tag_element.text unless tag_element.nil?
+
+      webview_url_element = project.at_xpath('webview-url')
+      @webview_url = default_webview_url
+      @webview_url = webview_url_element.text unless webview_url_element.nil?
 
       @exclude_pattern = []
       project.xpath('exclude-pattern').each do |ep|
         @exclude_pattern.push(ep.text)
+      end
+    end
+
+    # Generate the default webview url for the projects
+    # stored on github.
+    # For other projects return value is `connection`.
+    def default_webview_url
+      if @type.eql?('git') && @connection.include?('github.com')
+        "#{@connection}/tree/#{@tag}"
+      else
+        @connection
       end
     end
 
@@ -43,15 +57,24 @@ module PmdTester
     end
 
     def get_pmd_report_path(branch_name)
-      "#{get_project_target_dir(branch_name)}/pmd_report.xml"
+      if branch_name.nil?
+        nil
+      else
+        "#{get_project_target_dir(branch_name)}/pmd_report.xml"
+      end
     end
 
     def get_report_info_path(branch_name)
-      "#{get_project_target_dir(branch_name)}/report_info.json"
+      if branch_name.nil?
+        nil
+      else
+        "#{get_project_target_dir(branch_name)}/report_info.json"
+      end
     end
 
     def get_project_target_dir(branch_name)
-      dir = "target/reports/#{branch_name}/#{@name}"
+      branch_filename = PmdBranchDetail.branch_filename(branch_name)
+      dir = "target/reports/#{branch_filename}/#{@name}"
       FileUtils.mkdir_p(dir) unless File.directory?(dir)
       dir
     end