+++ test/test_diff_report_builder.rb
@@ -45,10 +45,10 @@ class TestDiffReportBuilder < Test::Unit::TestCase
     build_pmd_report(ORIGINAL_PATCH_PMD_REPORT_PATH, PATCH_PMD_REPORT_PATH)
 
     diff_builder = PmdTester::DiffBuilder.new
-    report_diffs = diff_builder.build(BASE_PMD_REPORT_PATH, PATCH_PMD_REPORT_PATH,
-                                      BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
+    project.report_diff = diff_builder.build(BASE_PMD_REPORT_PATH, PATCH_PMD_REPORT_PATH,
+                                             BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
 
-    PmdTester::DiffReportBuilder.new.build(project, report_diffs)
+    PmdTester::DiffReportBuilder.new.build(project)
 
     # Checking  css resources are copied into the diff report directory.
     assert_equal(true, File.exist?("#{css_path}/maven-base.css"))
@@ -64,8 +64,8 @@ class TestDiffReportBuilder < Test::Unit::TestCase
     project = PmdTester::ProjectsParser.new.parse('test/resources/project-list.xml')[1]
     actual_report_path = "target/reports/diff/#{project.name}"
 
-    report_diffs = PmdTester::ReportDiff.new
-    PmdTester::DiffReportBuilder.new.build(project, report_diffs)
+    project.report_diff = PmdTester::ReportDiff.new
+    PmdTester::DiffReportBuilder.new.build(project)
 
     # Checking the content of diff report is expected.
     expected_file = File.open(EXPECTED_EMPTY_REPORT_PATH).read